+++ src/assign-card/handler/assign-card.js
@@ -1,8 +1,8 @@
 const { batchEventMapper } = require('@ranty/nbased/handler')
 const inputMode = require('@ranty/nbased/handler/input/batchEventQueue')
 const outputMode = require('@ranty/nbased/handler/output/batchEventConfirmation')
-const assignGiftDomain = require('../domain/assignGift')
+const assignCardDomain = require('../domain/assign-card')
 
 module.exports.handler = async (events, context) => {
-  return batchEventMapper({ events, context }, inputMode, assignGiftDomain, outputMode)
+  return batchEventMapper({ events, context }, inputMode, assignCardDomain, outputMode)
 }
