--- src/assign-card/helper/calcAge.js
@@ -1,9 +1,0 @@
-async function calculateAge(birthday) {
-  // birthday is a string in format YYYYMMDD
-  const birthDate = new Date(birthday)
-  const ageDifMs = Date.now() - birthDate.getTime()
-  const ageDate = new Date(ageDifMs)
-  return Math.abs(ageDate.getUTCFullYear() - 1970)
-}
-
-module.exports = { calculateAge }
